requirements/dev.txt
pytest==2.4.2
pytest-cov==1.6
httpretty==0.7.0
furl==0.3.6
Flask-WebTest==0.0.4